calm-orbit-todo/phase2-3-fullstack/backend/requirements.txt
fastapi>=0.115.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
alembic>=1.14.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
python-decouple>=3.8
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0
openai-agents>=0.0.3
mcp[cli]>=1.2.0
httpx>=0.27.0
